requirements.txt
pillow
pyarrow
fonttools
pyyaml
tqdm
